Creating a Real Estate Advertisement Module (Chapter 8)
Besides completing chapter 8, this commit has changes relating to the correct formating and syntaxe of the file names and code, i.e. we used CamelCase for 'class EstateProperty(models.Model):' in the model files, we also corrected the '_name = 'estate.property'' by using '.' instead of '_' in the same file and the correspandant files in views, we also kept using the '_' in defining the models in the security/ir.model.access.csv file, those changed are applied on all models.